API created and deployed

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -5,22 +5,29 @@ using System.Text;
 using System.Threading.Tasks;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Entities;
+using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Respositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
 {
-    public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantRepository restaurantRepository) : IRequestHandler<DeleteRestaurantCommand, bool>
+    public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantRepository restaurantRepository, IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
     {
-        public async Task<bool> Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
+        public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
             logger.LogInformation($"Deleting restaurant with id: {request.Id}");
             var restaurant = await restaurantRepository.GetByIdAsync( request.Id );
             if ( restaurant == null )
             {
-                return false;
+                throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+            }
+            if(!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Delete))
+            {
+                throw new ForbidException();
             }
             await restaurantRepository.DeleteAsync(restaurant);
-            return true;
+            //return true;
         }
     }
 }
